Extend S4 Methods for seurat Class

Provide limited SingleCellExperiment-like method support.

# S4 method for seurat,ANY
assay(x, i, ...)

# S4 method for seurat
assayNames(x, ...)

# S4 method for seurat
assays(x, ..., withDimnames = TRUE)

# S4 method for seurat
barcodeRanksPerSample(object, ...)

# S4 method for seurat
cell2sample(object)

# S4 method for seurat
cellCountsPerCluster(object, interestingGroups)

# S4 method for seurat
clusterCellCountsPerSample(object)

# S4 method for seurat
colData(x, ...)

# S4 method for seurat,DataFrame
colData(x) <- value

# S4 method for seurat
colnames(x)

# S4 method for seurat
convertGenesToSymbols(object)

# S4 method for seurat
counts(object, ...)

# S4 method for seurat
diffExp(object, numerator, denominator,
  caller = c("edgeR", "DESeq2"), minCellsPerGene = 25L,
  minCountsPerCell = 5L)

# S4 method for seurat
fetchGeneData(object, genes, gene2symbol = FALSE)

# S4 method for seurat
fetchReducedDimData(object, reducedDim,
  dimsUse = c(1L, 2L), interestingGroups)

# S4 method for seurat
fetchReducedDimExpressionData(object, genes, reducedDim)

# S4 method for seurat
gene2symbol(object)

# S4 method for seurat
interestingGroups(object)

# S4 method for seurat,character
interestingGroups(object) <- value

# S4 method for seurat
metadata(x, ...)

# S4 method for seurat
metadata(x) <- value

# S4 method for seurat
metrics(object, ...)

# S4 method for seurat
metricsPerSample(object, f = c("mean", "median",
  "sum"))

# S4 method for seurat
plotBarcodeRanks(object, ...)

# S4 method for seurat
plotCellCounts(object, interestingGroups,
  fill = getOption("bcbio.discrete.fill", NULL), title = "cell counts")

# S4 method for seurat
plotCellTypesPerCluster(object, markers,
  reducedDim = c("TSNE", "UMAP"), expression = c("mean", "median",
  "sum"), headerLevel = 2L, ...)

# S4 method for seurat
plotDot(object, genes, colMin = -2.5, colMax = 2.5,
  dotMin = 0L, dotScale = 6L,
  color = getOption("bcbio.discrete.color", NULL),
  legend = getOption("bcbio.legend", TRUE))

# S4 method for seurat
plotFeature(object, features, reducedDim = c("TSNE",
  "UMAP"), color = getOption("bcbio.discrete.color", NULL),
  pointSize = getOption("bcbio.pointSize", 0.75),
  pointAlpha = getOption("bcbio.pointAlpha", 0.75),
  label = getOption("bcbio.label", TRUE),
  labelSize = getOption("bcbio.labelSize", 6L),
  dark = getOption("bcbio.dark", FALSE), grid = getOption("bcbio.grid",
  FALSE), legend = getOption("bcbio.legend", FALSE),
  aspectRatio = getOption("bcbio.aspectRatio", 1L))

# S4 method for seurat
plotGene(object, genes, geom = c("dot", "violin"),
  legend = getOption("bcbio.legend", TRUE))

# S4 method for seurat
plotGenesPerCell(object, geom = c("violin",
  "ridgeline", "ecdf", "histogram", "boxplot"), interestingGroups,
  min = 0L, max = Inf, trans = "log2",
  fill = getOption("bcbio.discrete.fill", NULL),
  title = "genes per cell")

# S4 method for seurat
plotKnownMarkersDetected(object, markers,
  reducedDim = c("TSNE", "UMAP"), headerLevel = 2L, ...)

# S4 method for seurat
plotMarker(object, genes, reducedDim = c("TSNE",
  "UMAP"), expression = c("mean", "median", "sum"),
  color = getOption("bcbio.discrete.color", NULL),
  pointSize = getOption("bcbio.pointSize", 0.75),
  pointAlpha = getOption("bcbio.pointAlpha", 0.8),
  pointsAsNumbers = FALSE, label = getOption("bcbio.label", TRUE),
  labelSize = getOption("bcbio.labelSize", 6L),
  dark = getOption("bcbio.dark", FALSE), grid = getOption("bcbio.grid",
  FALSE), legend = getOption("bcbio.legend", TRUE),
  aspectRatio = getOption("bcbio.aspectRatio", 1L), title = TRUE)

# S4 method for seurat
plotMitoRatio(object, geom = c("violin", "ridgeline",
  "ecdf", "histogram", "boxplot"), interestingGroups, max = 1L,
  fill = getOption("bcbio.discrete.fill", NULL), trans = "sqrt",
  title = "mito ratio")

# S4 method for seurat
plotMitoVsCoding(object, interestingGroups,
  trendline = FALSE, color = getOption("bcbio.discrete.color", NULL),
  trans = "log2", title = "mito vs. coding")

# S4 method for seurat
plotNovelty(object, geom = c("violin", "ridgeline",
  "ecdf", "histogram", "boxplot"), interestingGroups, min = 0L,
  fill = getOption("bcbio.discrete.fill", NULL), trans = "identity",
  title = "genes per UMI (novelty)")

# S4 method for seurat
plotPCA(object, ...)

# S4 method for seurat
plotQC(object, interestingGroups, geom = c("violin",
  "ridgeline", "ecdf", "histogram", "boxplot"), headerLevel = 2L,
  legend = getOption("bcbio.legend", FALSE), return = c("grid", "list",
  "markdown"))

# S4 method for seurat
plotReducedDim(object, reducedDim, dimsUse = c(1L,
  2L), interestingGroups = "ident",
  color = getOption("bcbio.discrete.color", NULL),
  pointSize = getOption("bcbio.pointSize", 0.75),
  pointAlpha = getOption("bcbio.pointAlpha", 0.75),
  pointsAsNumbers = FALSE, label = getOption("bcbio.label", TRUE),
  labelSize = getOption("bcbio.labelSize", 6L),
  dark = getOption("bcbio.dark", FALSE), grid = getOption("bcbio.grid",
  FALSE), legend = getOption("bcbio.legend", TRUE),
  aspectRatio = getOption("bcbio.aspectRatio", 1L), title = NULL)

# S4 method for seurat
plotTopMarkers(object, markers, n = 10L,
  direction = c("positive", "negative", "both"), coding = FALSE,
  reducedDim = c("TSNE", "UMAP"), headerLevel = 2L, ...)

# S4 method for seurat
plotTSNE(object, ...)

# S4 method for seurat
plotUMAP(object, ...)

# S4 method for seurat
plotUMIsPerCell(object, geom = c("violin",
  "ridgeline", "ecdf", "histogram", "boxplot"), interestingGroups,
  min = 0L, max = Inf, point = c("none", "inflection", "knee"),
  trans = "log10", color = getOption("bcbio.discrete.color", NULL),
  fill = getOption("bcbio.discrete.fill", NULL),
  title = "UMIs per cell")

# S4 method for seurat
plotUMIsVsGenes(object, interestingGroups,
  trendline = FALSE, color = getOption("bcbio.discrete.color", NULL),
  trans = "log2", title = "UMIs vs. genes")

# S4 method for seurat
plotViolin(object, genes, scale = c("count", "width",
  "area"), fill = getOption("bcbio.discrete.fill", NULL),
  legend = getOption("bcbio.legend", TRUE))

# S4 method for seurat
plotZerosVsDepth(object, interestingGroups,
  color = getOption("bcbio.discrete.color", NULL),
  title = "zeros vs. depth")

# S4 method for seurat
reducedDims(x)

# S4 method for seurat
rowData(x, ...)

# S4 method for seurat
rownames(x)

# S4 method for seurat
rowRanges(x)

# S4 method for seurat
sampleData(object, interestingGroups)

# S4 method for seurat,DataFrame
sampleData(object) <- value

# S4 method for seurat
sampleNames(object)

# S4 method for seurat
topBarcodes(object, n = 1000L,
  return = c("data.frame", "list"))

Arguments

x
Primary object.

...
Additional arguments.

object
Object.

interestingGroups
character or NULL. Character vector of interesting groups. Must be formatted in camel case and intersect with sampleData() colnames.

value
Object to assign.

genes
character. Gene identifiers. Must match the rownames of the object.

gene2symbol
data.frame. Gene-to-symbol mappings. Columns must contain geneID and geneName.

reducedDim
string. Name of reduced dimension matrix slotted in reducedDims(). Includes TNSE, UMAP, PCA, for example.

dimsUse
integer. Vector of length 2 that denotes the columns from the reduced dimension matrix to use for centerX and centerY column calculations. Defaults the first and second dimensions.

fill
ggproto/ScaleDiscrete or NULL. Desired ggplot2 fill scale. Must supply discrete values. When set to NULL, the default ggplot2 color palette will be used. If manual color definitions are desired, we recommend using ggplot2::scale_fill_manual(). To set the discrete fill palette globally, use options(bcbio.discrete.fill = scale_fill_viridis_d()).

title
string or NULL. Plot title.

expression
string. Calculation to apply. Uses match.arg() and defaults to the first argument in the character vector.

headerLevel
scalar integer (1-7). R Markdown header level.

color
ggproto/ScaleDiscrete or NULL. Desired ggplot2 color scale. Must supply discrete values. When set to NULL, the default ggplot2 color palette will be used. If manual color definitions are desired, we recommend using ggplot2::scale_color_manual(). To set the discrete color palette globally, use options(bcbio.discrete.color = scale_color_viridis_d()).

legend
boolean. Include plot legend.

pointSize
scalar numeric. Cell point size.

pointAlpha
scalar numeric (0-1). Alpha transparency level. Useful when there many cells in the dataset, and some cells can be masked.

label
boolean. Overlay a cluster identitiy label on the plot.

labelSize
scalar integer. Size of the text label.

dark
boolean. Plot against a dark background using basejump::theme_midnight().

grid
boolean. Show major grid lines but hide axis lines.

aspectRatio
scalar integer. Aspect ratio.

geom
string. Plot type. Uses match.arg() and defaults to the first argument in the character vector.

min
scalar numeric. Recommended minimum value cutoff.

max
scalar numeric. Recommended maximum value cutoff.

trans
string. Name of the axis scale transformation to apply. See help("scale_x_continuous", "ggplot2") for more information.

pointsAsNumbers
boolean. Plot the points as numbers (TRUE) or dots (FALSE).

trendline
boolean. Include trendline on plot.

return
string. Return type. Uses base::match.arg() internally and defaults to the first argument in the character vector.

Value

Match SummarizedExperiment method return.

See also